issue-6958: fastshard - extracted TFiberShard template from naive_mirrored to reuse it in production shard implementation
