+++ cloud/filestore/libs/storage/fastshard/impl/ya.make
@@ -6,6 +6,7 @@ RECURSE(
 IF (OPENSOURCE AND NOT FORCE_FASTSHARD_IPC_STUB)
     RECURSE(
         model
+        fiber_bridge
     )
 
     RECURSE_FOR_TESTS(
